价格为0 && 登录时效延长

+++ forms/rsp/u8.go
@@ -21,25 +21,25 @@ type LogList struct {
 }
 
 type LogDetail struct {
-	Id               int    `json:"id"`
-	UpdateTime       string `json:"update_time"`
-	PickId           int    `json:"pick_id"`
-	BatchId          int    `json:"batch_id"`
-	PrePickGoodsId   int    `json:"pre_pick_goods_id"`
-	OrderGoodsId     int    `json:"order_goods_id"`
-	Number           string `json:"number"`
-	ShopId           int    `json:"shop_id"`
-	DistributionType int    `json:"distribution_type"`
-	Sku              string `json:"sku"`
-	GoodsName        string `json:"goods_name"`
-	GoodsType        string `json:"goods_type"`
-	GoodsSpe         string `json:"goods_spe"`
-	Shelves          string `json:"shelves"`
-	DiscountPrice    int    `json:"discount_price"`
-	NeedNum          int    `json:"need_num"`
-	CompleteNum      int    `json:"complete_num"`
-	ReviewNum        int    `json:"review_num"`
-	Unit             string `json:"unit"`
+	Id               int     `json:"id"`
+	UpdateTime       string  `json:"update_time"`
+	PickId           int     `json:"pick_id"`
+	BatchId          int     `json:"batch_id"`
+	PrePickGoodsId   int     `json:"pre_pick_goods_id"`
+	OrderGoodsId     int     `json:"order_goods_id"`
+	Number           string  `json:"number"`
+	ShopId           int     `json:"shop_id"`
+	DistributionType int     `json:"distribution_type"`
+	Sku              string  `json:"sku"`
+	GoodsName        string  `json:"goods_name"`
+	GoodsType        string  `json:"goods_type"`
+	GoodsSpe         string  `json:"goods_spe"`
+	Shelves          string  `json:"shelves"`
+	DiscountPrice    float64 `json:"discount_price"`
+	NeedNum          int     `json:"need_num"`
+	CompleteNum      int     `json:"complete_num"`
+	ReviewNum        int     `json:"review_num"`
+	Unit             string  `json:"unit"`
 }
 
 type LogDetailRsp struct {
